pyproject.toml
[project]
name = "trading-agent"
version = "0.1.0"
description = "AI-powered stock investment decision system with multi-agent debate"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "孟钦学", email = "mengqinxue@users.noreply.github.com"}]

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "akshare>=1.12.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]